* commit '225a4af94f3faf5ca3726e86bc96cdda4c99a469': (12063 commits)
  Remove files from tests folder
  Wi-SUN FAN 1.1 dynamic MDR data request enabler
  Corrected BR removing of waiting list entry when supplicant is in key storage
  Adjust trace levels (#2692)
  Added reset for pan id and version to BR network start
  Renaming and cleaning ws bootstrap (#2688)
  Add Wi-SUN host configuration (#2690)
  Refactoring Wi-SUN stack (#2686)
  Renamed operation mode to operating mode.
  RF config resolver and some refactoring (#2683)
  Fixed WS IE PCAP read operation wrong length usage.
  Config: Remove additional HAVE_WS_ROUTER (#2684)
  Added API for configure supported Phy capability.
  Wi-SUN FAN 1.1 PCAP IE update
  Moved State machine and timer functions to own files
  Corrected system time check function return values
  Moved Wi-SUN Bootstrap Event handling to separate device handlers
  MAC to support mode switch on single channel (#2678)
  Added storing of PAN ID to NVM in BBR
  Corrected system time jump detection on BR startup
  ...

+++ nanostack/source/Service_Libs/fhss/channel_list.c
@@ -35,7 +35,7 @@ static bool channel_list_bit_test32(uint32_t word, int_fast8_t bit_number)
 {
     bool bitSet;
 
-    if (word & ((uint32_t) 1 << bit_number)) {
+    if (word & (1U << bit_number)) {
         bitSet = true;
     } else {
         bitSet = false;
